requirements.txt
pandas
numpy
scikit-learn
scipy
nltk
spacy
hdbscan
umap-learn
matplotlib
plotly
tqdm
pyyaml
sentence-transformers
pyarrow
